requirements.txt
# Core data processing
pandas==2.2.2
openpyxl==3.1.2
numpy==1.26.4

# Data warehouse
duckdb==0.10.3

# Dashboard
streamlit==1.35.0
plotly==5.22.0

# ML
scikit-learn==1.5.0
joblib==1.4.2

# GenAI / RAG
langchain==0.2.6
langchain-community==0.2.6
chromadb==0.5.0
sentence-transformers==3.0.1

# Utility
python-dotenv==1.0.1
schedule==1.2.2
